修改专辑和作品的一些bug

+++ Tumcca/Tumcca/src/com/pineapple/mobilecraft/tumcca/data/Album.java
@@ -23,7 +23,7 @@ public class Album {
      */
     public transient long id = -1;
 
-    public transient int author = -1;
+    public transient long author = -1;
 
     public transient List<WorksInfo> worksInfoList = null;
